FMC: remove unused and misleading conditionals
At registration and unregistration time, I was checking fmc->flags
for FMC_DEVICE_NO_MEZZANINE, to skip initialization and cleanup for
empty slots.  But the check was wrong (see previous commit) and
the conditional was never triggered.

I think we should register slots with no mezzanines, because the
carrier may offer some support anyways (the SPEC does), so the commit
removes the conditionals instead of fixing them.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>